# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PrintFlow API - Moran Creative",
        "description": "API para gestión de pedidos de impresión 3D",
        "contact": {
            "email": "admin@morancreative.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/admin/stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Estadísticas del dashboard (solo admin)",
                "operationId": "934bda1c5d4732ef612d3d51e9ef367e",
                "responses": {
                    "200": {
                        "description": "Estadísticas generales del taller",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orders_this_month": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "pending_review": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "in_production": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "ready_for_pickup": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "revenue_this_month": {
                                            "type": "number",
                                            "example": 8500
                                        },
                                        "active_clients": {
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "orders_by_status": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Solo admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/clients": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar clientes registrados (solo admin)",
                "operationId": "293f8d9df87a127e1dc8eadc87d6719e",
                "responses": {
                    "200": {
                        "description": "Lista de clientes con conteo de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "orders_count": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Solo admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}